+++ src/lib/components/table/proposals/StatusChip.tsx
@@ -28,9 +28,11 @@ const getBgColor = (
 
 export const StatusChip = ({
   status,
+  isTransparent = false,
   hasCloseBtn = false,
 }: {
   status: Proposal["status"];
+  isTransparent?: boolean;
   hasCloseBtn?: boolean;
 }) => {
   const formatStatus =
@@ -38,7 +40,11 @@ export const StatusChip = ({
       ? "Deposit Failed"
       : status.replace(/([A-Z])/g, " $1").trim();
   return (
-    <Tag bgColor={getBgColor(status)}>
+    <Tag
+      border="1px solid"
+      borderColor={getBgColor(status)}
+      bgColor={isTransparent ? "transparent" : getBgColor(status)}
+    >
       {formatStatus}
       {hasCloseBtn && <CustomIcon name="close" boxSize={3} mr={0} />}
     </Tag>